# web/openapi.json
{
  "components": {
    "schemas": {
      "AgentConfig": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "default": "",
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "default": "",
            "title": "Base Url",
            "type": "string"
          },
          "code": {
            "default": "",
            "title": "Code",
            "type": "string"
          },
          "context_window": {
            "default": 1000000,
            "minimum": 0.0,
            "title": "Context Window",
            "type": "integer"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "model": {
            "default": "",
            "title": "Model",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "use_responses": {
            "default": false,
            "title": "Use Responses",
            "type": "boolean"
          }
        },
        "title": "AgentConfig",
        "type": "object"
      },
...
      "AgentPoolConfig": {
        "additionalProperties": false,
        "properties": {
          "max_size": {
            "default": 256,
            "minimum": 1.0,
            "title": "Max Size",
            "type": "integer"
          },
          "sweep_interval_seconds": {
            "default": 60,
            "minimum": 1.0,
            "title": "Sweep Interval Seconds",
            "type": "integer"
          },
          "ttl_seconds": {
            "default": 1800,
            "minimum": 0.0,
            "title": "Ttl Seconds",
            "type": "integer"
          }
        },
        "title": "AgentPoolConfig",
        "type": "object"
      },
      "AgentRuntimeConfig": {
        "additionalProperties": false,
        "properties": {
          "context_budget_model_call_ratio": {
            "default": 0.8,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Context Budget Model Call Ratio",
            "type": "number"
          },
          "context_compression_enabled": {
            "default": true,
            "title": "Context Compression Enabled",
            "type": "boolean"
          },
          "context_compression_hard_stop_ratio": {
            "default": 0.98,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Context Compression Hard Stop Ratio",
            "type": "number"
          },
          "context_compression_min_items": {
            "default": 12,
            "minimum": 1.0,
            "title": "Context Compression Min Items",
            "type": "integer"
          },
          "context_compression_preserve_recent_items": {
            "default": 20,
            "minimum": 1.0,
            "title": "Context Compression Preserve Recent Items",
            "type": "integer"
          },
          "context_compression_preserve_recent_ratio": {
            "default": 0.25,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Context Compression Preserve Recent Ratio",
            "type": "number"
          },
          "context_compression_summary_max_tokens": {
            "default": 8000,
            "minimum": 512.0,
            "title": "Context Compression Summary Max Tokens",
            "type": "integer"
          },
          "context_compression_target_ratio": {
            "default": 0.2,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Context Compression Target Ratio",
            "type": "number"
          },
          "context_compression_trigger_ratio": {
            "default": 0.9,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Context Compression Trigger Ratio",
            "type": "number"
          },
          "main_max_turns": {
            "default": 1000,
            "minimum": 1.0,
            "title": "Main Max Turns",
            "type": "integer"
          },
          "model_stream_idle_timeout_seconds": {
            "default": 300,
            "minimum": 30.0,
            "title": "Model Stream Idle Timeout Seconds",
            "type": "integer"
          },
          "subordinate_max_turns": {
            "default": 1000,
            "minimum": 1.0,
            "title": "Subordinate Max Turns",
            "type": "integer"
          }
        },
        "title": "AgentRuntimeConfig",
        "type": "object"
      },
...
      "CommonResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[Any]",
        "type": "object"
      },
...
      "CommonResponse_DeleteSandboxImageResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeleteSandboxImageResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[DeleteSandboxImageResponse]",
        "type": "object"
      },
...
      "CommonResponse_InstanceConfigSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstanceConfigSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[InstanceConfigSchema]",
        "type": "object"
      },
...
      "CommonResponse_ListManagedHostImagesResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListManagedHostImagesResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[ListManagedHostImagesResponse]",
        "type": "object"
      },
...
      "CommonResponse_QuerySandboxImagesResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuerySandboxImagesResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[QuerySandboxImagesResponse]",
        "type": "object"
      },
...
      "CommonResponse_SandboxImageSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SandboxImageSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[SandboxImageSchema]",
        "type": "object"
      },
...
      "CommonResponse_UpdateInstanceConfigResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UpdateInstanceConfigResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[UpdateInstanceConfigResponse]",
        "type": "object"
      },
...
      "CreateSandboxImageRequest": {
        "properties": {
          "control_proxy_port": {
            "default": 8000,
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Control Port",
            "type": "integer"
          },
          "image_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Image Name",
            "type": "string"
          },
          "supports_tor": {
            "default": false,
            "title": "Supports Tor",
            "type": "boolean"
          }
        },
        "required": [
          "image_name"
        ],
        "title": "CreateSandboxImageRequest",
        "type": "object"
      },
...
      "DeleteSandboxImageResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "DeleteSandboxImageResponse",
        "type": "object"
      },
...
      "InstanceConfigSchema": {
        "additionalProperties": false,
        "properties": {
          "agent_pool": {
            "$ref": "#/components/schemas/AgentPoolConfig"
          },
          "agent_runtime": {
            "$ref": "#/components/schemas/AgentRuntimeConfig"
          },
          "agents": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AgentConfig"
            },
            "title": "Agents",
            "type": "object"
          }
        },
        "title": "InstanceConfigSchema",
        "type": "object"
      },
...
      "ListManagedHostImagesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ManagedHostImageSchema"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ListManagedHostImagesResponse",
        "type": "object"
      },
...
      "ManagedHostImageSchema": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "image_hash": {
            "default": "",
            "title": "Image Hash",
            "type": "string"
          },
          "image_id": {
            "default": "",
            "title": "Image Id",
            "type": "string"
          },
          "image_name": {
            "title": "Image Name",
            "type": "string"
          },
          "image_size": {
            "default": 0,
            "format": "int64",
            "title": "Image Size",
            "type": "integer"
          }
        },
        "required": [
          "image_name"
        ],
        "title": "ManagedHostImageSchema",
        "type": "object"
      },
...
      "QuerySandboxImagesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SandboxImageSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "items"
        ],
        "title": "QuerySandboxImagesResponse",
        "type": "object"
      },
...
      "SandboxImageSchema": {
        "properties": {
          "control_proxy_port": {
            "title": "Control Port",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_name": {
            "title": "Image Name",
            "type": "string"
          },
          "supports_tor": {
            "title": "Supports Tor",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "image_name",
          "control_proxy_port",
          "supports_tor",
          "created_at",
          "updated_at"
        ],
        "title": "SandboxImageSchema",
        "type": "object"
      },
...
      "UpdateAgentConfigRequest": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "context_window": {
            "minimum": 0.0,
            "title": "Context Window",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "title": "Name",
            "type": "string"
          },
          "use_responses": {
            "title": "Use Responses",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "description",
          "base_url",
          "api_key",
          "model",
          "use_responses",
          "context_window"
        ],
        "title": "UpdateAgentConfigRequest",
        "type": "object"
      },
...
      "UpdateInstanceConfigRequest": {
        "additionalProperties": false,
        "properties": {
          "agent_pool": {
            "$ref": "#/components/schemas/AgentPoolConfig"
          },
          "agent_runtime": {
            "$ref": "#/components/schemas/AgentRuntimeConfig"
          },
          "agents": {
            "additionalProperties": {
              "$ref": "#/components/schemas/UpdateAgentConfigRequest"
            },
            "title": "Agents",
            "type": "object"
          }
        },
        "required": [
          "agents",
          "agent_pool",
          "agent_runtime"
        ],
        "title": "UpdateInstanceConfigRequest",
        "type": "object"
      },
      "UpdateInstanceConfigResponse": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/InstanceConfigSchema"
          },
          "restarted": {
            "title": "Restarted",
            "type": "boolean"
          }
        },
        "required": [
          "config",
          "restarted"
        ],
        "title": "UpdateInstanceConfigResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AccessTokenAuth": {
        "in": "header",
        "name": "X-Z3r0-Access-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Z3r0 - A controlled multi-agent workbench for authorized security assessment, code auditing, internal review, and controlled research.",
    "version": "0.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/hosts/{id}/images": {
      "get": {
        "operationId": "list_managed_host_images_handler_api_hosts__id__images_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_ListManagedHostImagesResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Managed host not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "List Managed Host Images Handler",
        "tags": [
          "hosts"
        ]
      }
    },
...
    "/api/sandbox-images": {
      "get": {
        "operationId": "query_sandbox_images_route_api_sandbox_images_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QuerySandboxImagesResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Query Sandbox Images Route",
        "tags": [
          "sandbox-images"
        ]
      },
      "post": {
        "operationId": "create_sandbox_image_handler_api_sandbox_images_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_SandboxImageSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Create Sandbox Image Handler",
        "tags": [
          "sandbox-images"
        ]
      }
    },
    "/api/sandbox-images/{id}": {
      "delete": {
        "operationId": "delete_sandbox_image_handler_api_sandbox_images__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_DeleteSandboxImageResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Sandbox image not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Delete Sandbox Image Handler",
        "tags": [
          "sandbox-images"
        ]
      }
    },
    "/api/system-config/instance": {
      "get": {
        "operationId": "get_instance_config_handler_api_system_config_instance_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_InstanceConfigSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Get Instance Config Handler",
        "tags": [
          "system-config"
        ]
      },
      "patch": {
        "operationId": "update_instance_config_handler_api_system_config_instance_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstanceConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_UpdateInstanceConfigResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Update Instance Config Handler",
        "tags": [
          "system-config"
        ]
      }
    },
...
  }
}